refactor: colocate gatewayResourceName with its handler per review
